Fixed metavars and added a parser
Now metavariables will not unify with with constants that are not in
scope where the metavariable was created.

Also a nice parser is in place and you can test it with:

Main.hs> testIt "dist/test.tex" "∃x(P(x) -> ∀y P(y))"

Isn't that lovely?

->       -- implication
| \/     -- or
& /\     -- and
~ !      -- not
∀ forall -- forall
∃ exists -- exists

Predicates starts with upper case letter and varialbe with lower.
Functions are written as lower e.g P(f(x))